feat(arm32): M4-6 完成 —— PCB/TCB 结构体,并把汇编边界收窄成"指向 ctx 的指针"
## 字段怎么来的:三份清单(§4.7)

`arch/tcb.h` 里**每个字段后面都写着它在 include/task/pcb.h 里的行号**,
那不是装饰,是可以拿源文件逐行核对的账。三类:

  ARCH_INDEP     逐字段沿用 x86 的名字(parent_group/thread_queue/file_open/
                 brk_current/eevdf_vruntime…),类型按 ARM 取
  ARCH_DEP       整份 PCB 里只有 pagedir 一个;TCB 里是 ctx 与 vfp
  暂缓           逐条列表 + 汇总成 ARM_TCB_DEFERRED_FIELDS 常量,便于将来 grep

pcb_t/tcb_t 是指针 typedef(pcb.h:46-47),这条约定照抄。

## ★ 一处刻意的偏离:一个任务只要一个内核栈 ★

x86 每个 TCB 有两个 1MB 内核栈,原因很具体:**x86 的 syscall 指令不会切栈**,
handler.S:315 必须从 TCB 里把 syscall_stack 装进 %rsp。
而中断走 TSS.rsp0,于是天然是两段栈。

ARM 的 SVC 与 IRQ 各有**硬件 banked** 的 SP,进异常时自动切,
不需要从任何结构体里读栈指针 ⇒ 留 syscall_stack 等于每任务白留 1MB。

连带:M4-5 的栈池口径从"32 栈 = 16 任务"变成"32 栈 = 32 任务"(池大小不用改)。

## ★★ 最值得记的一条:汇编的边界是"指向 ctx 的指针" ★★

第一版照 x86 的样子,在 taskctx_asm.h 里定义 ARM_TCB_OFF_CTX 等四个偏移,
让汇编按偏移访问 TCB —— **做不下去,而且不该那么做**:

1. TCB 有大量指针字段(宿主 8 字节 / 目标 4 字节),于是**宿主机上算出来的
   偏移与目标板不同**,那几个 _Static_assert 在宿主上验的是错的布局
   (与 heap_block_t 的 sizeof 同一个坑);
2. 更要紧的是**汇编根本不需要它** ——

     x86:syscall 不切栈 -> 必须从 TCB 读 syscall_stack(handler.S:315)
     ARM:硬件 banked SP,自动切
     x86:per-CPU 当前任务按 %gs:0x4c0 取(而那个偏移一条断言都没有)
     ARM:TPIDRPRW 拿结构体指针,再按偏移取 current_task
     x86:切换器就地改写 C 栈上的 registers_t
     ARM:C 侧把 &tcb->ctx 传进汇编 —— 边界上走的是**指针**不是偏移

⇒ ARM 侧汇编要钉的偏移只剩两组(arm_task_ctx_t 内部、percpu_t.current_task),
而且**两组都只含 u32 字段**,宿主与目标布局一致 ——
**断言在宿主机上验的就是目标板的真实偏移**。这比 x86 少一层,是架构给的。

顺带把 percpu_t 的地址字段从 uintptr_t 改成 u32(ARM 上一样 4 字节),
唯一理由是:结构体里只要有一个指针宽度的字段,"在宿主机上验偏移"就做不成。
改完 sizeof(percpu_t) == 44 在两个平台上一致。

## 一个被断言逼出来的真问题:VFP 区对齐

最初写 aligned(32)("对齐足一点稳"),而 heap_alloc 只保证 8 字节 ——
构成未定义行为。真实要求是 **4 字节**(ARMv7-A 的 VFP 传输规则);
x86 的 16 是 fxsave64 的硬件要求,**ARM 没有对应物**。
改掉之后加了一条**故意的绊线**:

    _Static_assert(_Alignof(tcb) == 8u, "有新字段要求更强的对齐,而 heap_alloc 只保证 8");

将来谁照抄 x86 的 aligned(16),编译期就会被迫面对分配器的问题。

## 汇编侧

- boot/context.S:arch_ctx_layout_check() 按宏读回 17 个字段,
  用**返回码**而不是"打印后停机" —— 布局错了该报 FAIL 后继续跑
- 一组 .if/.error 静态断言,钉住**指令序列所依赖的前提**
  (r[] 必须在 ctx 开头、槽是 4 字节步长)—— 这些在 C 侧看不出来

## 验收

  verify_board.py --load   **44 passed / 0 failed**(新增 tcb_ctx_layout)
  板上实测                 Tcb ctx : sizeof=512 align_ok=1 -> PASS (result=0)
  guard_trip.py / 其余     无回归
  pytest tests/            9 failed / 72 passed(9 项预先存在)

## 顺带修掉一个我自己弄坏的工具

tmp-test/spurious_probe.py 里 PERCPU_SIZE 是硬编码的 40,
加了 current_task 之后过期了 —— 它会安静地去读 CPU1 结构体错位后的字节,
读出来的数看着像模像样、其实毫无意义。改成每次让宿主编译器现算。

+++ arch/arm32/src/percpu_hw.c
@@ -7,7 +7,7 @@
 #include <arch/cpu.h>
 #include <arch/percpu.h>
 
-percpu_t *percpu_init_self(uintptr_t stack_top)
+percpu_t *percpu_init_self(u32 stack_top)
 {
     u32       mpidr = arch_read_mpidr();
     u32       id    = mpidr & 0x3u;
